refactor(gateway): 移动AuthCenter头文件并修复base64解码变量类型
将AuthCenter头文件从Gateway.cpp移到Gateway.h以优化依赖关系
将base64Decode函数中的in_变量类型从int改为size_t以避免潜在的类型转换问题

+++ src/gateway/AuthCenter.cpp
@@ -330,7 +330,7 @@ std::string base64Decode(const std::string& input) {
     std::string result;
     int i = 0;
     int j = 0;
-    int in_ = 0;
+    size_t in_ = 0;
     unsigned char char_array_4[4], char_array_3[3];
     
     while (str.size() > in_ && str[in_] != '=') {
